Refactor - Created EoW Data for EOW Data Layer on map
