+++ src/app/app.component.ts
@@ -1,33 +1,30 @@
 import {Component, OnInit, Inject, AfterViewInit} from '@angular/core';
 import {DOCUMENT} from '@angular/common';
-import debounce from 'lodash/debounce';
 import Map from 'ol/Map';
 import OSM from 'ol/source/OSM';
 import TileLayer from 'ol/layer/Tile';
 import View from 'ol/View';
-import VectorSource from 'ol/source/Vector';
-import VectorLayer from 'ol/layer/Vector';
 import {fromLonLat} from 'ol/proj';
 
-import GeoJSON from 'ol/format/GeoJSON';
-import CircleStyle from 'ol/style/Circle';
-import {
-  Style,
-  Stroke,
-  Fill
-} from 'ol/style';
-
-import {
-  printStats,
-  calculateStats,
-} from './utils';
-import colors from './colors.json';
+// import CircleStyle from 'ol/style/Circle';
+// import {
+//   Style,
+//   Stroke,
+//   Fill
+// } from 'ol/style';
+//
+// import {
+//   printStats,
+//   calculateStats,
+// } from './utils';
+// import colors from './colors.json';
 import {HttpClient} from '@angular/common/http';
 import {PieChart} from './pie-chart';
 import {Popup} from './popup';
 import {Layers} from './layers';
 import {MeasurementStore} from './measurement-store';
 import {UserStore} from './users';
+import {EowData} from './eow-data';
 
 @Component({
   selector: 'app-root',
@@ -40,6 +37,7 @@ export class AppComponent implements OnInit {
   popupObject: any;
   measurementStore: MeasurementStore;
   userStore: UserStore;
+  eowData: EowData;
   dataLayer: any;
   allDataSource: any;
   pieChart: any;
@@ -54,6 +52,7 @@ export class AppComponent implements OnInit {
     this.popupObject = new Popup(this.document, this.pieChart, this.userStore);
     this.layers = new Layers(this.document, this.http);
     this.measurementStore = new MeasurementStore();
+    this.eowData = new EowData();
   }
 
   ngOnInit() {
@@ -62,55 +61,56 @@ export class AppComponent implements OnInit {
     this.measurementStore.init(this.map, this.dataLayer, this.allDataSource);
     this.layers.addLayers(this.map);
     this.userStore.init();
+    this.eowData.init(this.map, this.htmlDocument, this.userStore, this.measurementStore);
 
     this.setupEventHandlers();
   }
 
   private initMap() {
-    const WFS_URL = 'https://geoservice.maris.nl/wms/project/eyeonwater_australia?service=WFS'
-      + '&version=1.0.0&request=GetFeature&typeName=eow_australia&maxFeatures=5000&outputFormat=application%2Fjson';
-    const styleCache = {};
+    // const WFS_URL = 'https://geoservice.maris.nl/wms/project/eyeonwater_australia?service=WFS'
+    //   + '&version=1.0.0&request=GetFeature&typeName=eow_australia&maxFeatures=5000&outputFormat=application%2Fjson';
+    // const styleCache = {};
 
-    this.allDataSource = new VectorSource({
-      format: new GeoJSON(),
-      url: WFS_URL
-    });
+    // this.allDataSource = new VectorSource({
+    //   format: new GeoJSON(),
+    //   url: WFS_URL
+    // });
 
 // Style Features using ..... FU values (called for each feature on every render call)
-    const basicStyle = (feature, resolution) => {
-      const fuValue = feature.get('fu_value');
-      const styleKey = `${fuValue}_${resolution}`;
-      // Avoid some unnecessary computation
-      if (styleCache[styleKey]) {
-        return styleCache[styleKey];
-      }
-      feature.set('visible', true);
-      const styleOptions = {
-        image: new CircleStyle({
-          radius: this.map.getView().getZoom() * Math.log2(5),
-          stroke: new Stroke({
-            color: 'white'
-          }),
-          fill: new Fill({
-            color: colors[fuValue]
-          })
-        })
-      };
-
-      styleCache[styleKey] = new Style(styleOptions);
-      return styleCache[styleKey];
-    };
-    this.dataLayer = new VectorLayer({
-      source: this.allDataSource,
-      style: basicStyle
-    });
-    this.dataLayer.set('name', 'EOW Data');
-
-    this.dataLayer.on('change', debounce(({target}) => {
-      // Populate datalayer
-      const element = this.document.querySelector('.sub-header-stats') as HTMLElement;
-      element.innerHTML = printStats(calculateStats(target.getSource().getFeatures()), this.userStore);
-    }, 200));
+//     const basicStyle = (feature, resolution) => {
+//       const fuValue = feature.get('fu_value');
+//       const styleKey = `${fuValue}_${resolution}`;
+//       // Avoid some unnecessary computation
+//       if (styleCache[styleKey]) {
+//         return styleCache[styleKey];
+//       }
+//       feature.set('visible', true);
+//       const styleOptions = {
+//         image: new CircleStyle({
+//           radius: this.map.getView().getZoom() * Math.log2(5),
+//           stroke: new Stroke({
+//             color: 'white'
+//           }),
+//           fill: new Fill({
+//             color: colors[fuValue]
+//           })
+//         })
+//       };
+//
+//       styleCache[styleKey] = new Style(styleOptions);
+//       return styleCache[styleKey];
+//     };
+//     this.dataLayer = new VectorLayer({
+//       source: this.allDataSource,
+//       style: basicStyle
+//     });
+//     this.dataLayer.set('name', 'EOW Data');
+//
+//     this.dataLayer.on('change', debounce(({target}) => {
+//       // Populate datalayer
+//       const element = this.document.querySelector('.sub-header-stats') as HTMLElement;
+//       element.innerHTML = printStats(calculateStats(target.getSource().getFeatures()), this.userStore);
+//     }, 200));
     const mainMap = new TileLayer({
       source: new OSM()
     });
@@ -120,7 +120,7 @@ export class AppComponent implements OnInit {
       target: 'map',
       layers: [
         mainMap,
-        this.dataLayer
+        // this.dataLayer
       ],
       view: new View({
         center: fromLonLat([133.945313, -26.431228]),
@@ -189,8 +189,6 @@ export class AppComponent implements OnInit {
     this.document.getElementById('clearFilterButton').addEventListener('click', (event) => {
       this.clearFilter();
     });
-
-    this.allDataSource.on('change', this.measurementStore.initialLoadMeasurements.bind(this.measurementStore));
   }
 
   private clearFilter() {
