+++ src/app/eow-data.spec.ts
@@ -1,0 +1,7 @@
+import { EowData } from './eow-data';
+
+describe('EowData', () => {
+  it('should create an instance', () => {
+    expect(new EowData()).toBeTruthy();
+  });
+});
